--- a/package.json
+++ b/package.json
@@ -18,7 +18,7 @@
     "clean": "rm -rf public/packs/*",
     "prebuild": "yarn run clean",
     "build": "yarn build:js",
-    "build:js": "concurrently 'yarn:webpack' 'yarn:generate-browsers-json'",
+    "build:js": "concurrently 'yarn:webpack' 'make browsers.json'",
     "build:css": "build-sass app/assets/stylesheets/*.css.scss app/components/*.scss --load-path=app/assets/stylesheets --out-dir=app/assets/builds"
   },
   "dependencies": {